+++ app/Http/Controllers/Admin/DirectoryController.php
@@ -10,6 +10,20 @@ use Yajra\DataTables\Facades\DataTables;
 
 class DirectoryController extends Controller
 {
+    public $user;
+    
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = auth()->guard('web')->user();
+            return $next($request);
+        });
+
+        $this->middleware('permission:view-directory')->only('index');
+        $this->middleware('permission:create-directory')->only(['create', 'store']);
+        $this->middleware('permission:edit-directory')->only(['edit', 'update']);
+        $this->middleware('permission:delete-directory')->only('destroy');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -24,18 +38,28 @@ class DirectoryController extends Controller
                     $editUrl = route('directory.edit', $category->id);
                     $deleteUrl = route('directory.destroy', $category->id);
 
-                    return '
-                        <button type="button" class="btn btn-sm btn-primary text-white edit_btn editBtn" data-id="' . $category->id . '">
-                            <i class="fa fa-edit"></i>
-                        </button>
+                    $buttons = '';
 
-                        <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $category->id . '" style="display:inline-block;">
-                            ' . csrf_field() . method_field('DELETE') . '
-                            <button type="submit" class="btn btn-danger btn-sm deleteItem">
-                                <i class="fa fa-trash"></i>
+                    if (auth()->user()->can('edit-directory')) {
+                        $buttons .= '
+                            <button type="button" class="btn btn-sm btn-primary text-white edit_btn editBtn" data-id="' . $category->id . '">
+                                <i class="fa fa-edit"></i>
                             </button>
-                        </form>
-                    ';
+                        ';
+                    }
+
+                    if (auth()->user()->can('delete-directory')) {
+                        $buttons .= '
+                            <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $category->id . '" style="display:inline-block;">
+                                ' . csrf_field() . method_field('DELETE') . '
+                                <button type="submit" class="btn btn-danger btn-sm deleteItem">
+                                    <i class="fa fa-trash"></i>
+                                </button>
+                            </form>
+                        ';
+                    }
+
+                    return $buttons ?: '<span class="text-muted">No Actions</span>';
                 })
                 ->rawColumns(['status', 'action'])
                 ->make(true);
