+++ app/Http/Controllers/Admin/GalleryController.php
@@ -10,6 +10,20 @@ use Yajra\DataTables\Facades\DataTables;
 
 class GalleryController extends Controller
 {
+    public $user;
+    
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = auth()->guard('web')->user();
+            return $next($request);
+        });
+
+        $this->middleware('permission:view-gallery')->only('index');
+        $this->middleware('permission:create-gallery')->only(['create', 'store']);
+        $this->middleware('permission:edit-gallery')->only(['edit', 'update']);
+        $this->middleware('permission:delete-gallery')->only('destroy');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -38,20 +52,29 @@ class GalleryController extends Controller
                     $editUrl = route('gallery.edit', $photo->id);
                     $deleteUrl = route('gallery.destroy', $photo->id);
 
-                    return '
-                        <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
-                            <i class="fa fa-edit"></i>
-                        </a>
-                        <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $photo->id . '" style="display:inline-block;">
-                            ' . csrf_field() . method_field('DELETE') . '
-                            <button type="submit" class="btn btn-danger btn-sm deleteItem">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        </form>
-                    ';
+                    $buttons = '';
+                    if (auth()->user()->can('edit-gallery')) {
+                        $buttons .= '
+                            <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
+                                <i class="fa fa-edit"></i>
+                            </a>
+                        ';
+                    }
+                    if (auth()->user()->can('delete-gallery')) {
+                        $buttons .= '
+                            <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $photo->id . '" style="display:inline-block;">
+                                ' . csrf_field() . method_field('DELETE') . '
+                                <button type="submit" class="btn btn-danger btn-sm deleteItem">
+                                    <i class="fa fa-trash"></i>
+                                </button>
+                            </form>
+                        ';
+                    }
+
+                    return $buttons ?: '<span class="text-muted">No Actions</span>';
                 })
 
-                ->rawColumns(['image', 'action']) // prevent escaping HTML
+                ->rawColumns(['image', 'action']) 
                 ->make(true);
         }
         return view('backend.gallery.index');
