+++ app/Http/Controllers/Admin/WidgetController.php
@@ -10,6 +10,20 @@ use Yajra\DataTables\Facades\DataTables;
 
 class WidgetController extends Controller
 {
+    public $user;
+    
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = auth()->guard('web')->user();
+            return $next($request);
+        });
+
+        $this->middleware('permission:view-widget')->only('index');
+        $this->middleware('permission:create-widget')->only(['create', 'store']);
+        $this->middleware('permission:edit-widget')->only(['edit', 'update']);
+        $this->middleware('permission:delete-widget')->only('destroy');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -44,17 +58,28 @@ class WidgetController extends Controller
                     $editUrl = route('widget.edit', $widget->id);
                     $deleteUrl = route('widget.destroy', $widget->id);
 
-                    return '
-                        <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
-                            <i class="fa fa-edit"></i>
-                        </a>
-                        <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $widget->id . '" style="display:inline-block;">
-                            ' . csrf_field() . method_field('DELETE') . '
-                            <button type="submit" class="btn btn-danger btn-sm deleteItem">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        </form>
-                    ';
+                    $buttons = '';
+
+                    if (auth()->user()->can('edit-widget')) {
+                        $buttons .= '
+                            <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
+                                <i class="fa fa-edit"></i>
+                            </a>
+                        ';
+                    }
+
+                    if (auth()->user()->can('delete-widget')) {
+                        $buttons .= '
+                            <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $widget->id . '" style="display:inline-block;">
+                                ' . csrf_field() . method_field('DELETE') . '
+                                <button type="submit" class="btn btn-danger btn-sm deleteItem">
+                                    <i class="fa fa-trash"></i>
+                                </button>
+                            </form>
+                        ';
+                    }
+
+                    return $buttons ?: '<span class="text-muted">No Actions</span>';
                 })
                 ->filterColumn('bn_title', function($query, $keyword) {
                     $query->whereHas('translations', function($q) use ($keyword) {
